backend/requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
pydantic
python-multipart
anthropic
python-dotenv
pytest
faker
